Uploaded the Google OAuth API
Most of this is in the web app files, except vendor/google/auth and possibly some others I haven't seen you because I'm stuck on the auth error.

Figured I'd update it here just so at least our GitHub is complete in case we need to demo through localhost or something.